Successfuly retrieved urls for every house in Medellin
This is a huge step. We can replicate this with every city. The whole scraper is almost complete. Only drawback is that the URLcollector.py file is somewhat slow. It may take a total of 1h to retrieve the links of 2000 pages (that's an aproximate of the number of pages in the whole site).